Claim details if no LAA ref
Amendments to claim details page to account for if there is not an LAA reference when NSM supplemental/appeal/amendment is selected. Note that this code needs tidying up when I have more time - there's a lot of repeated code that I believe can be cleaned up but the logic works in the proto for now.