Updates to routes if no LAA ref
Amendments to the proto routing to account for if there is not an LAA reference when NSM supplemental/appeal/amendment is selected. Proto should ask for LAA ref if user has selected Yes to laa ref, and proto should skip ref and display a form is user has select No to laa ref.