Update laa-reference.html
Changing the design and flow of the payment request.

Removing the question 'Is there an LAA reference for the original claim?' and the autocomplete suggestions box for finding the LAA reference for the original claim. Replacing this with a new screen called 'Search for the original claim' which uses an inline search to allow users to either find and link a claim, or continue without linking and create a new digital record.