Update claim-details.html
Updating if statements to use the new 'linkedCRM' data to decide what fields to show. 'Yes' means that we have an LAA reference for the original claim and it will be linked to the payment request and we can pre-populate some fields. 'No' means that there isn't a reference to the original claim so we will require more details from the user in this form.