Payments v1-4 (nsm-1-4)
Versioning the payments work to keep a record of our old design.

I've created 'Payments v1-4' to save the previous payments design which includes the 'Is there an LAA ref' question and the autocomplete component.

In the proto files, this work is now saved under version 'nsm-1-4'

Going forward, new designs will now be added to version 'latest-portal'