Redesign of the office code autocomplete component
Removing the autocomplete component used to captured the office account code and the firm name as this component is not technically feasible.

Replaced with a new screen to capture the office code and then a new screen to allow the caseworker to confirm that the code and the firm name matches their records. Also added a new screen if the office code is invalid and not found in our system.

Updated routes and laa-reference pages to route to date received page before the claim details page, and removed date received and office code from the claim details page as they are now captured on the new pages.