Updating LAA references used in payment requests
Updating laa-reference.html to set the LAA references that will be used throughout the journey. Now the proto should assign laaRef 1-4 to the 'laa-ref' variable when 'Select' button is clicked so that it can be used on pages like check-answers and submit-confirmation.

Have also updated the claim-summary to show the chosen reference for the linked AC claim.